+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
   (`Issue #644 <https://github.com/adamchainz/django-mysql/issues/644>`__)
 * Added ``JSONIn`` lookup which only works with literal values (not with
   expressions nor subqueries).
+* Fix ``JSONContains`` to make it work with scalar values again.
+  (`PR #668 <https://github.com/adamchainz/django-mysql/pull/668>`__).
 
 3.5.0 (2020-05-04)
 ------------------